pyproject.toml
[tool.poetry]
name = "escola-api"
version = "0.1.0"
description = ""
authors = ["lucianogomes02 <luciano@pontotel.com.br>"]
readme = "README.md"
packages = [{include = "escola_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
black = "^23.1.0"
validate-docbr = "^1.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
